requirements.txt
# LangChain core
langchain>=0.3.0
langchain-classic>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.0
langchain-chroma>=0.2.0

# Embeddings
sentence-transformers>=3.0.0

# Vector store
chromadb>=0.5.0

# LLM
groq>=0.11.0

# Document loaders
PyPDF2>=3.0.0
python-docx>=1.1.0
openpyxl>=3.1.0

# News API
requests>=2.31.0

# Web Server
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9
httpx>=0.27.0
itsdangerous>=2.1.0

# Environment
python-dotenv>=1.0.0

# Auth (JWT + password hashing + Google OAuth)
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5
authlib>=1.3.0

# Tests
pytest>=8.0.0

# Tests
pytest>=8.0.0
